LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/bailout-side-effects-ignore-opens-during-unload.window-expected.txt


PASS document.open bailout should not have any side effects (ignore-opens-during-unload is greater than 0 during unload event)
PASS document.open bailout should not have any side effects (ignore-opens-during-unload is greater than 0 during beforeunload event)
PASS document.open bailout should not have any side effects (ignore-opens-during-unload is greater than 0 during pagehide event)
